cumulative reference update after http://mercurial.mcs.anl.gov/ad/OpenADFortTk/rev/f922e5f97043  and http://mercurial.mcs.anl.gov/ad/OpenAnalysis/rev/492b56563606

+++ testRoundTrip/examples/Loop_simple_1f/refOutput/admjointhead_sf.pre.xb.x2w.w2f.post.f
@@ -108,7 +108,7 @@ C original function
 C$OPENAD XXX Template ad_template.f
 C$OPENAD XXX Simple loop
       DO I = 0, 3, 1
-        Y(INT(I+1))%v = X(I+1)%v
+        Y(I+1)%v = X(I+1)%v
       END DO
 
 C original function end
@@ -125,7 +125,7 @@ C taping
 C$OPENAD XXX Template ad_template.f
 C$OPENAD XXX Simple loop
       DO I = 0,3,1
-        Y(INT(I+1))%v = X(I+1)%v
+        Y(I+1)%v = X(I+1)%v
       END DO
 
 C taping end
