+++ testRoundTrip/examples/Loop_simple_1f/refOutput/tlmnonehead_sf.pre.xb.x2w.w2f.post.f
@@ -49,7 +49,7 @@ C
 C$OPENAD XXX Template ad_template.f
 C$OPENAD XXX Simple loop
       DO I = 0, 3, 1
-        Y(INT(I+1))%v = X(I+1)%v
+        Y(I+1)%v = X(I+1)%v
         CALL setderiv(Y(I+1),X(I+1))
       END DO
       END SUBROUTINE
