+++ testRoundTrip/examples/boxmodel/refOutput/admsplithead_sf.pre.xb.x2w.w2f.post.f
@@ -288,7 +288,7 @@ C$OPENAD XXX Template ad_template.f
       END DO
       DO L = 1, 3, 1
         TSVEC(INT(L)) = TNOW(L)%v
-        TSVEC(INT(L+3)) = SNOW(L)%v
+        TSVEC(L+3) = SNOW(L)%v
       END DO
           end if
           if (our_rev_mode%tape) then
@@ -304,7 +304,7 @@ C$OPENAD XXX Template ad_template.f
       OpenAD_Symbol_41 = 0_w2f__i8
       DO L = 1,3,1
         TSVEC(INT(L)) = TNOW(L)%v
-        TSVEC(INT(L+3)) = SNOW(L)%v
+        TSVEC(L+3) = SNOW(L)%v
         OpenAD_Symbol_41 = (INT(OpenAD_Symbol_41)+INT(1_w2f__i8))
       END DO
       integer_tape(integer_tape_pointer) = OpenAD_Symbol_41
